+++ CHANGELOG.md
@@ -25,6 +25,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Fixed parsing of attachment's filename from header within `MailgunMailer`. remp/remp#1386
   - Previous implementation incorrectly parsed filenames with dash. Filename of attached file "invoice-2024-09-24.pdf" would be only last part "24.pdf".
   - Added `MailHeaderTrait` with method `getHeaderParameter()` and tests to validate it.
+- Fixed mail type stats when groupped by week or month. remp/remp#1374
 
 ## Archive